Fixed bug, abstract class to UI buttons
- Changed name of ControlHandlerObject prefab to Game Manager.
- Fixed Null Exception bug.
- Added abstract class to UI Buttons to make it more modular.
