+++ CuddleTreeDefenseGame/Assets/Scripts/UI/UIButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 40526671b8db4d84dbc39f5c077b1253
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
